Cleanup and rewrite to utilize "plugin" statefulness across testsuite. Not yet implemented the testclass lifecycle, using simple "object" as its still works as intended, might need to re-implement

+++ plugin/inject-solr-plugin/src/main/java/org/hrodberaht/injection/plugin/junit/solr/SolrTestRunner.java
@@ -239,8 +239,10 @@ public class SolrTestRunner {
 
     private void tearDown() throws IOException {
         File coreDir = new File(home, coreName);
-        LOG.info("cleaning : "+coreDir.getAbsolutePath());
-        FileUtils.cleanDirectory(coreDir);
+        if(coreDir.exists()) {
+            LOG.info("cleaning : " + coreDir.getAbsolutePath());
+            FileUtils.cleanDirectory(coreDir);
+        }
     }
